Some additions in activitypoints workers and motd settings
added a setting to give a bonuses for vip users in activitypoints worker.
added a setting in motd_text to send a link in motd.

corrected a bug:
When a room is deleted, the navigator does not load the user's room list.
Now when a room is deleted the navigator loads the room list correctly.