Update on Gifts and Pets
When a gift is placed in the room it will update to the package that was chosen by the sender.

Now when a pet is scratched they will win experience points.

Update on catalogue:
Added Coco items & catalogue page.
Added new items definitions to Pixels Exchange furnis.
Changed behaviour definition of present furnis to Gift.